feat: refactor tech detection + add detection for React
