+++ package.json
@@ -44,7 +44,9 @@
         "openai": "^5.8.2",
         "path": "^0.12.7",
         "prettier": "^3.7.4",
-        "puppeteer": "^24.34.0",
+        "puppeteer": "^24.43.1",
+        "puppeteer-extra": "^3.3.6",
+        "puppeteer-extra-plugin-stealth": "^2.11.2",
         "readline": "^1.3.0",
         "ses": "^1.13.0",
         "yaml": "^2.8.0",
@@ -62,7 +64,7 @@
         "@types/babel__traverse": "^7.28.0",
         "@types/better-sqlite3": "^7.6.13",
         "@types/blessed": "^0.1.25",
-        "@types/node": "^24.5.2",
+        "@types/node": "^24.12.4",
         "ts-node": "^10.9.2",
         "tsconfig-paths": "^4.2.0",
         "typescript": "^5.9.2"
