+++ src/lazyLoad/techDetect/checkNuxtJS.ts
@@ -1,0 +1,33 @@
+import * as cheerio from "cheerio";
+
+/**
+ * Detects if a webpage uses Nuxt.js by checking for "/_nuxt" paths in src or href attributes.
+ *
+ * @param $ - The Cheerio API object containing the parsed HTML
+ * @returns Promise that resolves to an object with detection status and evidence
+ */
+export const checkNuxtJS = async ($: cheerio.CheerioAPI) => {
+    let detected = false;
+    let evidence = "";
+
+    // go through the page source, and check for "/_nuxt" in the src or href attribute
+    $("*").each((_, el) => {
+        // @ts-ignore
+        const tag = $(el).get(0).tagName;
+        // @ts-ignore
+        const attribs = el.attribs;
+        if (attribs) {
+            for (const [attrName, attrValue] of Object.entries(attribs)) {
+                if (attrName === "src" || attrName === "href") {
+                    // @ts-ignore
+                    if (attrValue.includes("/_nuxt")) {
+                        detected = true;
+                        evidence = `${attrName} :: ${attrValue}`;
+                    }
+                }
+            }
+        }
+    });
+
+    return { detected, evidence };
+};
